feat: Refactor da aplicação do euComida para Pedido
