+++ pedido-service/pom.xml
@@ -2,10 +2,10 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.eucomida</groupId>
+    <groupId>br.com.fabiohaider</groupId>
     <artifactId>pedido-service</artifactId>
     <name>pedido-service</name>
-    <description>Serviço de pedidos do euComida</description>
+    <description>Serviço de pedidos do pedido</description>
     <version>1.0.3-SNAPSHOT</version>
     <packaging>jar</packaging>
