+++ pedido-service/src/main/java/br/com/fabiohaider/pedido/PedidoServiceApplication.java
@@ -1,0 +1,12 @@
+package br.com.fabiohaider.pedido;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PedidoServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(PedidoServiceApplication.class, args);
+    }
+}
