+++ pedido-service/src/main/java/br/com/fabiohaider/pedido/domain/enums/Status.java
@@ -1,0 +1,9 @@
+
+package br.com.fabiohaider.pedido.domain.enums;
+
+@SuppressWarnings("unused")
+public enum Status {
+    EM_ANDAMENTO,
+    ENTREGUE,
+    CANCELADO
+}
