+++ pedido-service/src/main/java/br/com/fabiohaider/pedido/infrastructure/config/MicrometerConfig.java
@@ -1,0 +1,16 @@
+package br.com.fabiohaider.pedido.infrastructure.config;
+
+import io.micrometer.core.instrument.binder.system.ProcessorMetrics;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@SuppressWarnings("unused")
+public class MicrometerConfig {
+
+    @Bean
+    public ProcessorMetrics processorMetrics() {
+        return new ProcessorMetrics();
+    }
+
+}
