+++ pedido-service/src/main/java/br/com/fabiohaider/pedido/infrastructure/controller/PedidoController.java
@@ -1,0 +1,39 @@
+package br.com.fabiohaider.pedido.infrastructure.controller;
+
+import br.com.fabiohaider.pedido.application.ports.in.ConsultarStatusPedidoUseCase;
+import br.com.fabiohaider.pedido.application.ports.in.CriarPedidoUseCase;
+import br.com.fabiohaider.pedido.domain.enums.Status;
+import br.com.fabiohaider.pedido.domain.model.Pedido;
+import br.com.fabiohaider.pedido.infrastructure.controller.dto.PedidoRequest;
+import jakarta.validation.Valid;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.UUID;
+
+@RestController
+@RequestMapping("/pedidos")
+@SuppressWarnings("unused")
+public class PedidoController {
+
+    private final CriarPedidoUseCase criarPedido;
+    private final ConsultarStatusPedidoUseCase consultarStatus;
+
+    public PedidoController(CriarPedidoUseCase criarPedido, ConsultarStatusPedidoUseCase consultarStatus) {
+        this.criarPedido = criarPedido;
+        this.consultarStatus = consultarStatus;
+    }
+
+    @ResponseStatus(HttpStatus.CREATED)
+    @PostMapping
+    public Pedido criarPedido(@Valid @RequestBody PedidoRequest request) throws RuntimeException {
+        return criarPedido.criar(request.clienteId());
+    }
+
+    @ResponseStatus(HttpStatus.OK)
+    @GetMapping("/{id}/status")
+    public Status statusPedido(@PathVariable UUID id) {
+        return consultarStatus.consultar(id);
+    }
+
+}
