+++ pedido-service/src/main/java/br/com/fabiohaider/pedido/infrastructure/controller/dto/PedidoRequest.java
@@ -1,0 +1,7 @@
+package br.com.fabiohaider.pedido.infrastructure.controller.dto;
+
+import jakarta.validation.constraints.NotNull;
+import java.util.UUID;
+
+public record PedidoRequest(@NotNull UUID clienteId) {
+}
