--- pedido-service/src/main/java/com/eucomida/pedido/PedidoServiceApplication.java
@@ -1,12 +1,0 @@
-package com.eucomida.pedido;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PedidoServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(PedidoServiceApplication.class, args);
-    }
-}
