--- pedido-service/src/main/java/com/eucomida/pedido/application/ports/in/ConsultarStatusPedidoUseCase.java
@@ -1,10 +1,0 @@
-
-package com.eucomida.pedido.application.ports.in;
-
-import com.eucomida.pedido.domain.enums.Status;
-import java.util.UUID;
-
-@FunctionalInterface
-public interface ConsultarStatusPedidoUseCase {
-    Status consultar(UUID pedidoId);
-}
