--- pedido-service/src/main/java/com/eucomida/pedido/application/ports/in/CriarPedidoUseCase.java
@@ -1,10 +1,0 @@
-
-package com.eucomida.pedido.application.ports.in;
-
-import com.eucomida.pedido.domain.model.Pedido;
-import java.util.UUID;
-
-@FunctionalInterface
-public interface CriarPedidoUseCase {
-    Pedido criar(UUID clienteId);
-}
