--- pedido-service/src/main/java/com/eucomida/pedido/domain/enums/Status.java
@@ -1,9 +1,0 @@
-
-package com.eucomida.pedido.domain.enums;
-
-@SuppressWarnings("unused")
-public enum Status {
-    EM_ANDAMENTO,
-    ENTREGUE,
-    CANCELADO
-}
