--- pedido-service/src/main/java/com/eucomida/pedido/infrastructure/config/MicrometerConfig.java
@@ -1,16 +1,0 @@
-package com.eucomida.pedido.infrastructure.config;
-
-import io.micrometer.core.instrument.binder.system.ProcessorMetrics;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@SuppressWarnings("unused")
-public class MicrometerConfig {
-
-    @Bean
-    public ProcessorMetrics processorMetrics() {
-        return new ProcessorMetrics();
-    }
-
-}
