--- pedido-service/src/main/java/com/eucomida/pedido/infrastructure/controller/dto/PedidoRequest.java
@@ -1,7 +1,0 @@
-package com.eucomida.pedido.infrastructure.controller.dto;
-
-import jakarta.validation.constraints.NotNull;
-import java.util.UUID;
-
-public record PedidoRequest(@NotNull UUID clienteId) {
-}
