--- pedido-service/src/main/java/com/eucomida/pedido/infrastructure/exception/GlobalExceptionHandler.java
@@ -1,18 +1,0 @@
-package com.eucomida.pedido.infrastructure.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-@RestControllerAdvice
-@SuppressWarnings("unused")
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(RuntimeException.class)
-    public ResponseEntity<String> handleRuntime(RuntimeException ex) {
-        if (ex.getMessage().contains("Pedido não encontrado")) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ex.getMessage());
-        }
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(ex.getMessage());
-    }
-}
