--- pedido-service/src/main/java/com/eucomida/pedido/infrastructure/persistence/repository/SpringJpaRepository.java
@@ -1,10 +1,0 @@
-
-package com.eucomida.pedido.infrastructure.persistence.repository;
-
-import com.eucomida.pedido.infrastructure.persistence.entity.PedidoEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface SpringJpaRepository extends JpaRepository<PedidoEntity, UUID> {
-}
