+++ pedido-service/src/main/resources/application-docker.properties
@@ -1,4 +1,4 @@
-spring.datasource.url=jdbc:postgresql://localhost:5432/eucomida
+spring.datasource.url=jdbc:postgresql://localhost:5432/pedido
 spring.datasource.username=postgres
 spring.datasource.password=postgres
 spring.datasource.driver-class-name=org.postgresql.Driver
@@ -7,10 +7,10 @@ spring.jpa.hibernate.ddl-auto=update
 spring.jpa.open-in-view=false
 spring.jpa.show-sql=false
 
-spring.security.oauth2.resourceserver.jwt.issuer-uri=http://localhost:8080/realms/eucomida
+spring.security.oauth2.resourceserver.jwt.issuer-uri=http://localhost:8080/realms/pedido
 
 logging.level.root=INFO
-logging.level.com.eucomida=DEBUG
+logging.level.com.pedido=DEBUG
 
 management.endpoints.web.exposure.include=*
 management.endpoint.prometheus.enabled=true
