--- pedido-service/src/test/java/com/eucomida/pedido/config/TestSecurityConfig.java
@@ -1,21 +1,0 @@
-package com.eucomida.pedido.config;
-
-import org.mockito.Mockito;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Profile;
-import org.springframework.security.oauth2.jwt.JwtDecoder;
-import org.springframework.security.oauth2.jwt.NimbusJwtDecoder;
-
-import javax.crypto.SecretKey;
-import javax.crypto.spec.SecretKeySpec;
-
-@TestConfiguration
-@Profile("test")
-public class TestSecurityConfig {
-
-    @Bean
-    public JwtDecoder jwtDecoder() {
-        return Mockito.mock(JwtDecoder.class);
-    }
-}
